+++ enterprise/internal/campaigns/resolvers/permissions_test.go
@@ -20,9 +20,11 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/campaigns"
+	"github.com/sourcegraph/sourcegraph/internal/db"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/github"
 	"github.com/sourcegraph/sourcegraph/internal/repoupdater"
 )
@@ -248,6 +250,200 @@ func TestPermissionLevels(t *testing.T) {
 			}
 		})
 
+		t.Run("CampaignsCodeHosts", func(t *testing.T) {
+			tests := []struct {
+				name        string
+				currentUser int32
+				user        int32
+				wantErr     bool
+			}{
+				{
+					name:        "site-admin viewing other user",
+					currentUser: adminID,
+					user:        userID,
+					wantErr:     false,
+				},
+				{
+					name:        "non-site-admin viewing other's hosts",
+					currentUser: userID,
+					user:        adminID,
+					wantErr:     true,
+				},
+				{
+					name:        "non-site-admin viewing own hosts",
+					currentUser: userID,
+					user:        userID,
+					wantErr:     false,
+				},
+			}
+
+			for _, tc := range tests {
+				t.Run(tc.name, func(t *testing.T) {
+					pruneUserCredentials(t)
+
+					graphqlID := string(graphqlbackend.MarshalUserID(tc.user))
+
+					var res struct{ Node apitest.User }
+
+					input := map[string]interface{}{"user": graphqlID}
+					queryCodeHosts := `
+				  query($user: ID!) {
+				    node(id: $user) { ... on User { campaignsCodeHosts { totalCount } } }
+				  }
+                `
+
+					actorCtx := actor.WithActor(ctx, actor.FromUser(tc.currentUser))
+					errors := apitest.Exec(actorCtx, t, s, input, &res, queryCodeHosts)
+					if !tc.wantErr && len(errors) != 0 {
+						t.Fatal("got error but didn't expect one")
+					} else if tc.wantErr && len(errors) == 0 {
+						t.Fatal("expected error but got none")
+					}
+				})
+			}
+		})
+
+		t.Run("CampaignsCredentialByID", func(t *testing.T) {
+			tests := []struct {
+				name        string
+				currentUser int32
+				user        int32
+				wantErr     bool
+			}{
+				{
+					name:        "site-admin viewing other user",
+					currentUser: adminID,
+					user:        userID,
+					wantErr:     false,
+				},
+				{
+					name:        "non-site-admin viewing other's credential",
+					currentUser: userID,
+					user:        adminID,
+					wantErr:     true,
+				},
+				{
+					name:        "non-site-admin viewing own credential",
+					currentUser: userID,
+					user:        userID,
+					wantErr:     false,
+				},
+			}
+
+			for _, tc := range tests {
+				t.Run(tc.name, func(t *testing.T) {
+					pruneUserCredentials(t)
+
+					cred, err := db.UserCredentials.Create(ctx, db.UserCredentialScope{
+						Domain:              db.UserCredentialDomainCampaigns,
+						ExternalServiceID:   "https://github.com/",
+						ExternalServiceType: extsvc.TypeGitHub,
+						UserID:              tc.user,
+					}, &auth.OAuthBearerToken{Token: "SOSECRET"})
+					if err != nil {
+						t.Fatal(err)
+					}
+					graphqlID := string(marshalCampaignsCredentialID(cred.ID))
+
+					var res struct{ Node apitest.CampaignsCredential }
+
+					input := map[string]interface{}{"id": graphqlID}
+					queryCodeHosts := `
+				  query($id: ID!) {
+				    node(id: $id) { ... on CampaignsCredential { id } }
+				  }
+                `
+
+					actorCtx := actor.WithActor(ctx, actor.FromUser(tc.currentUser))
+					errors := apitest.Exec(actorCtx, t, s, input, &res, queryCodeHosts)
+					if !tc.wantErr && len(errors) != 0 {
+						t.Fatal("got error but didn't expect one")
+					} else if tc.wantErr && len(errors) == 0 {
+						t.Fatal("expected error but got none")
+					}
+					if !tc.wantErr {
+						if have, want := res.Node.ID, graphqlID; have != want {
+							t.Fatalf("invalid node returned, wanted ID=%q, have=%q", want, have)
+						}
+					}
+				})
+			}
+		})
+
+		t.Run("DeleteCampaignsCredential", func(t *testing.T) {
+			tests := []struct {
+				name        string
+				currentUser int32
+				user        int32
+				wantAuthErr bool
+			}{
+				{
+					name:        "site-admin for other user",
+					currentUser: adminID,
+					user:        userID,
+					wantAuthErr: false,
+				},
+				{
+					name:        "non-site-admin for other user",
+					currentUser: userID,
+					user:        adminID,
+					wantAuthErr: true,
+				},
+				{
+					name:        "non-site-admin for self",
+					currentUser: userID,
+					user:        userID,
+					wantAuthErr: false,
+				},
+			}
+
+			for _, tc := range tests {
+				t.Run(tc.name, func(t *testing.T) {
+					pruneUserCredentials(t)
+
+					cred, err := db.UserCredentials.Create(ctx, db.UserCredentialScope{
+						Domain:              db.UserCredentialDomainCampaigns,
+						ExternalServiceID:   "https://github.com/",
+						ExternalServiceType: extsvc.TypeGitHub,
+						UserID:              tc.user,
+					}, &auth.OAuthBearerToken{Token: "SOSECRET"})
+					if err != nil {
+						t.Fatal(err)
+					}
+
+					var res struct{ Node apitest.CampaignsCredential }
+
+					input := map[string]interface{}{
+						"campaignsCredential": marshalCampaignsCredentialID(cred.ID),
+					}
+					mutationDeleteCampaignsCredential := `
+					mutation($campaignsCredential: ID!) {
+						deleteCampaignsCredential(campaignsCredential: $campaignsCredential) { alwaysNil }
+					}
+                `
+
+					actorCtx := actor.WithActor(ctx, actor.FromUser(tc.currentUser))
+					errors := apitest.Exec(actorCtx, t, s, input, &res, mutationDeleteCampaignsCredential)
+					if tc.wantAuthErr {
+						if len(errors) != 1 {
+							t.Fatalf("expected 1 error, but got %d: %s", len(errors), errors)
+						}
+						if !strings.Contains(errors[0].Error(), "must be authenticated") {
+							t.Fatalf("wrong error: %s %T", errors[0], errors[0])
+						}
+					} else {
+						// We don't care about other errors, we only want to
+						// check that we didn't get an auth error.
+						for _, e := range errors {
+							if strings.Contains(e.Error(), "must be authenticated") {
+								t.Fatalf("auth error wrongly returned: %s %T", errors[0], errors[0])
+							}
+						}
+					}
+				})
+			}
+		})
+
 		t.Run("Campaigns", func(t *testing.T) {
 			tests := []struct {
 				name                string
@@ -322,7 +518,7 @@ func TestPermissionLevels(t *testing.T) {
 		})
 	})
 
-	t.Run("mutations", func(t *testing.T) {
+	t.Run("campaign mutations", func(t *testing.T) {
 		mutations := []struct {
 			name           string
 			mutationFunc   func(campaignID, changesetID, campaignSpecID string) string
@@ -465,9 +661,7 @@ func TestPermissionLevels(t *testing.T) {
 		}
 	})
 
-	t.Run("admin-only create mutations", func(t *testing.T) {
-		// These can be removed once we enable creation of
-		// changesetSpecs/campaignSpecs/applyCampaign for non-site-admin users.
+	t.Run("spec mutations", func(t *testing.T) {
 		mutations := []struct {
 			name         string
 			mutationFunc func(userID string) string
@@ -498,23 +692,23 @@ func TestPermissionLevels(t *testing.T) {
 					currentUser int32
 					wantAuthErr bool
 				}{
-					{
-						name:        "authorized user",
-						currentUser: userID,
-						wantAuthErr: true,
-					},
-					{
-						name:        "authorized site-admin",
-						currentUser: adminID,
-						wantAuthErr: false,
-					},
+					{name: "no user", currentUser: 0, wantAuthErr: true},
+					{name: "user", currentUser: userID, wantAuthErr: false},
+					{name: "site-admin", currentUser: adminID, wantAuthErr: false},
 				}
 
 				for _, tc := range tests {
 					t.Run(tc.name, func(t *testing.T) {
 						cleanUpCampaigns(t, store)
 
-						mutation := m.mutationFunc(string(graphqlbackend.MarshalUserID(tc.currentUser)))
+						namespaceID := string(graphqlbackend.MarshalUserID(tc.currentUser))
+						if tc.currentUser == 0 {
+							// If we don't have a currentUser we try to create
+							// a campaign in another namespace, solely for the
+							// purposes of this test.
+							namespaceID = string(graphqlbackend.MarshalUserID(userID))
+						}
+						mutation := m.mutationFunc(namespaceID)
 
 						actorCtx := actor.WithActor(ctx, actor.FromUser(tc.currentUser))
 
@@ -525,7 +719,7 @@ func TestPermissionLevels(t *testing.T) {
 							if len(errs) != 1 {
 								t.Fatalf("expected 1 error, but got %d: %s", len(errs), errs)
 							}
-							if !strings.Contains(errs[0].Error(), "must be site admin") {
+							if !strings.Contains(errs[0].Error(), "not authenticated") {
 								t.Fatalf("wrong error: %s %T", errs[0], errs[0])
 							}
 						} else {