+++ test/qa/test.sh
@@ -38,6 +38,7 @@ source /root/.profile
 pushd client/web || exit
 yarn run test:regression:core
 yarn run test:regression:integrations
+yarn run test:regression:search
 popd || exit
 PID=$(pgrep ffmpeg)
 kill "$PID"